Abstract away SearchBar
